Improve code and view models.

+++ GeolocationAds/TemplateViewModel/CaptureTemplateViewModel.cs
@@ -51,7 +51,7 @@ namespace GeolocationAds.TemplateViewModel
 
         public CaptureTemplateViewModel(Capture capture, ICaptureService service)
         {
-            this.ContentTypesTemplate = new ObservableCollection<ContentTypeTemplateViewModel>();
+            //this.ContentTypesTemplate = new ObservableCollection<ContentTypeTemplateViewModel>();
 
             this.RemoveCommand = new Command<Capture>(Remove);
 
@@ -61,7 +61,7 @@ namespace GeolocationAds.TemplateViewModel
 
             this.Service = service;
 
-            Task.Run(async () => { await FillTemplate(); });
+            //Task.Run(async () => { await FillTemplate(); });
         }
 
         public async Task FillTemplate()
@@ -70,22 +70,6 @@ namespace GeolocationAds.TemplateViewModel
             {
                 foreach (var item in this.CurrentAdvertisement.Contents)
                 {
-                    //if (item.Type == ContentVisualType.Image)
-                    //{
-                    //    var _template = ContentTypeTemplateFactory.BuilContentType(item, item.Content);
-
-                    //    this.ContentTypesTemplate.Add(_template);
-                    //}
-
-                    //if (item.Type == ContentVisualType.Video)
-                    //{
-                    //    var _file = await CommonsTool.SaveByteArrayToTempFile(item.Content);
-
-                    //    var _template = ContentTypeTemplateFactory.BuilContentType(item, _file);
-
-                    //    this.ContentTypesTemplate.Add(_template);
-                    //}
-
                     var _template = await AppToolCommon.ProcessContentItem(item);
 
                     this.ContentTypesTemplate.Add(_template);
@@ -93,18 +77,31 @@ namespace GeolocationAds.TemplateViewModel
             }
         }
 
+        public async Task InitializeAsync()
+        {
+            this.ContentTypesTemplate = new ObservableCollection<ContentTypeTemplateViewModel>();
+
+            await FillTemplate();
+        }
+
         public async void Remove(Capture capture)
         {
             try
             {
-                var _apiResponse = await this.Service.Remove(capture.ID);
+                var _removeResponse = await Shell.Current.DisplayAlert("Notification", $"Are you sure you want to remove this item?", "Yes", "No");
 
-                if (_apiResponse.IsSuccess)
+                if (_removeResponse)
                 {
-                    OnDeleteItem(EventArgs.Empty);
-                }
-                else
-                {
+                    var _apiResponse = await this.Service.Remove(capture.ID);
+
+                    if (_apiResponse.IsSuccess)
+                    {
+                        OnDeleteItem(EventArgs.Empty);
+                    }
+                    else
+                    {
+                        await CommonsTool.DisplayAlert("Error", _apiResponse.Message);
+                    }
                 }
             }
             catch (Exception ex)
